Remove Asumes.NotNull on SizeReady
In several cases, the m_inProgress.UnsafeData is indeed null. However, the method doesn't fail, and I presume it gets cast to an empty byte[]. Since the UnsafeOffset is 0 and the size is 0, nothing really happens.

I have run the Blueye app in debug mode, where this assertion is enforced, and I have seen the warning but no issues with the app.